Update default promotion dates to December 31, 2025
The countdown was showing 'offer expired' because the default date
was set to November 27, 2025 (Black Friday), which is now in the past.

Updated defaults:
- Top Bar Banner: 'HOLIDAY SALE' with   emojis
- Countdown: 'HOLIDAY SALE ENDS SOON!' ending Dec 31, 2025
- Changed default background_style to 'teal'

Note: If the old date is still showing, go to Admin Panel  Website
Promotions & Banners  Countdown Promo and update the End Date to a
future date, then click 'Save All Changes'.